+++ NEWS.md
@@ -3,6 +3,8 @@
 * `DESCRIPTION` version bumped to 0.3.1.
 * Can deactivate CDN modules available from [https://api.highcharts.com/highcharts](https://api.highcharts.com/highcharts) see (#2)
 * Fixed Shiny app GUI for downloading
+* Move arg `xlab` and `ylab` are related to figure presentation and not
+data specification. Move them to `hd_opts()` (#5) 
 
 # highdir 0.3.0